ci(changesets): versioning packages (#57)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @fuels/react@0.16.0

### Minor Changes

- 🐞 fix: fuel hooks will only re-render tracked properties, instead of
listening to every useQuery property, by
[@helciofranco](https://github.com/helciofranco) (See
[#55](https://github.com/FuelLabs/fuels-npm-packs/pull/55))

## @fuels/assets@0.16.0



## @fuels/changeset@0.16.0



## @fuels/eslint-plugin@0.16.0



## @fuels/jest@0.16.0



## @fuels/local-storage@0.16.0



## @fuels/prettier-config@0.16.0



## @fuels/react-xstore@0.16.0



## @fuels/playwright-utils@0.16.0



## @fuels/ts-config@0.16.0



## @fuels/tsup-config@0.16.0

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>