ci(changesets): versioning packages (#64)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @fuels/react@0.18.0

### Minor Changes

- 🐞 fix: only set theme colors style when its on client, by
[@LuizAsFight](https://github.com/LuizAsFight) (See
[#63](https://github.com/FuelLabs/fuels-npm-packs/pull/63))

## @fuels/assets@0.18.0



## @fuels/changeset@0.18.0



## @fuels/eslint-plugin@0.18.0



## @fuels/jest@0.18.0



## @fuels/local-storage@0.18.0



## @fuels/prettier-config@0.18.0



## @fuels/react-xstore@0.18.0



## @fuels/playwright-utils@0.18.0



## @fuels/ts-config@0.18.0



## @fuels/tsup-config@0.18.0

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>