ci(changesets): versioning packages (#69)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @fuels/react@0.18.1

### Patch Changes

- 🐞 fix: UI connectors loader, by
[@luizstacio](https://github.com/luizstacio) (See
[#68](https://github.com/FuelLabs/fuels-npm-packs/pull/68))

## @fuels/assets@0.18.1



## @fuels/changeset@0.18.1



## @fuels/eslint-plugin@0.18.1



## @fuels/jest@0.18.1



## @fuels/local-storage@0.18.1



## @fuels/prettier-config@0.18.1



## @fuels/react-xstore@0.18.1



## @fuels/playwright-utils@0.18.1



## @fuels/ts-config@0.18.1



## @fuels/tsup-config@0.18.1

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>