ci(changesets): versioning packages (#115)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @fuels/kms-account@0.24.0

### Minor Changes

- Create KMSAccount for using KMS from AWS with fuels typescript SDK, by
[@luizstacio](https://github.com/luizstacio) (See
[#120](https://github.com/FuelLabs/fuels-npm-packs/pull/120))

## @fuels/changeset@0.24.0



## @fuels/eslint-plugin@0.24.0



## @fuels/jest@0.24.0



## @fuels/local-storage@0.24.0



## @fuels/prettier-config@0.24.0



## @fuels/react-xstore@0.24.0



## @fuels/ts-config@0.24.0



## @fuels/tsup-config@0.24.0

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>