ci(changesets): versioning packages (#139)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @fuels/local-storage@0.26.0

### Minor Changes

- Check for localStorage availability, by
[@nelitow](https://github.com/nelitow) (See
[#132](https://github.com/FuelLabs/fuels-npm-packs/pull/132))

## @fuels/react-xstore@0.26.0

### Minor Changes

- Update tests to handle the new local-storage checks, by
[@nelitow](https://github.com/nelitow) (See
[#132](https://github.com/FuelLabs/fuels-npm-packs/pull/132))

## @fuels/changeset@0.26.0



## @fuels/eslint-plugin@0.26.0



## @fuels/jest@0.26.0



## @fuels/kms-account@0.26.0



## @fuels/prettier-config@0.26.0



## @fuels/proxy-cli@0.26.0



## @fuels/ts-config@0.26.0



## @fuels/tsup-config@0.26.0

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>